refactor: api 엔드포인트 통일 및 주석 인코딩 이슈 해결
