+++ .gitignore
@@ -37,3 +37,5 @@ out/
 .vscode/
 
 # 배포용 설정 파일 release 위해 .gitignore에서 삭제(25.02.17)
+
+.env
