+++ src/main/java/com/untitled/cherrymap/security/JoinController.java
@@ -1,20 +1,21 @@
-package com.untitled.cherrymap.controller.auth;
+package com.untitled.cherrymap.security;
 
-import com.untitled.cherrymap.dto.auth.JoinDTO;
+import com.untitled.cherrymap.security.dto.JoinDTO;
 import com.untitled.cherrymap.repository.MemberRepository;
-import com.untitled.cherrymap.service.auth.JoinService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Map;
 
 @RestController
+@RequestMapping("/api")
 @RequiredArgsConstructor
 @CrossOrigin
 public class JoinController {
