+++ src/main/java/com/untitled/cherrymap/security/dto/CustomUserDetailsDTO.java
@@ -1,4 +1,4 @@
-package com.untitled.cherrymap.dto.auth;
+package com.untitled.cherrymap.security.dto;
 
 import com.untitled.cherrymap.domain.Member;
 import lombok.Getter;
