+++ src/main/java/com/untitled/cherrymap/security/dto/JoinDTO.java
@@ -1,4 +1,4 @@
-package com.untitled.cherrymap.dto.auth;
+package com.untitled.cherrymap.security.dto;
 
 import lombok.Getter;
