+++ src/main/resources/application.properties
@@ -1,3 +1,5 @@
+spring.config.import=optional:file:.env[.properties]
+
 # Spring Boot ?????? ??
 spring.application.name=cherrymap
 
@@ -26,6 +28,6 @@ springdoc.swagger-ui.path=/cherrymap-ui.html
 springdoc.swagger-ui.tags-sorter=alpha
 springdoc.swagger-ui.operations-sorter=alpha
 
-spring.jwt.secret=vmfhaltmskdlstkfkdgodyroqkfwkdbalroqkfwkdbalaaaaaaaaaaaaaaaabbbbb
+spring.jwt.secret=${JWT_SECRET}
 
 server.port=8086
